Fix: array comparison breaks if using the environments plugin (#682)
* update comments from #249

I found the logic behind some of what happened here hard to follow and updated the comments to try to make it easier to follow.

* fix environments updating global array

The environments plugin was changing `MergeDeep.NAME_FIELDS`, which is a global object. The reason was to avoid environments being filtered out from the change list if they only have a name field.
However, the environments plugin has it's own overriden sync method, and thus we can simply drop the whole filtering from that method.

Fixes #108

* remove repeating line

This has to be a "typo" from 9a74e05 calling the same assignment twice. Removing to clean up.